requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
gevent==24.2.1
gevent-eventemitter==2.1
greenlet==3.0.3
idna==3.7
protobuf==3.20.3
pycryptodomex==3.20.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
steam==1.4.4
steamclient==1.2
urllib3==2.2.1
vdf==3.4
zope.event==5.0
zope.interface==6.4.post2
